# apps/wiki/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plunk API",
    "description": "Open-source email platform API for transactional emails, campaigns, and marketing automation",
    "version": "1.0.0",
    "contact": {
      "name": "Plunk Support",
      "url": "https://www.useplunk.com"
    }
  },
  "servers": [
    {
      "url": "https://next-api.useplunk.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key authentication. The project is automatically derived from the key.\n\n**`/v1/track` requires a public key (`pk_*`)** — it is the one endpoint intended for client-side use, and a secret key is rejected there with `401`.\n\n**Every other endpoint requires a secret key (`sk_*`)** and rejects public keys with `401`.\n\nSo the two key types are not interchangeable in either direction: pick the key that matches the endpoint you are calling."
      }
    },
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Dot-path of the offending field, e.g. `attachments.0.filename`."
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Validation issue code, e.g. `invalid_type`, `too_big`, `reserved_event`."
          },
          "received": {
            "description": "The value that was received, when available."
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code, e.g. `VALIDATION_ERROR`, `INVALID_API_KEY`, `IDEMPOTENCY_KEY_REUSED`."
              },
              "message": {
                "type": "string"
              },
              "statusCode": {
                "type": "integer"
              },
              "requestId": {
                "type": "string",
                "description": "Correlation ID for this request. Include it when contacting support."
              },
              "errors": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "description": "Field-level detail, present on validation failures."
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "description": "Additional error context."
              },
              "suggestion": {
                "type": "string",
                "description": "Hint for fixing the request."
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 255
        },
        "description": "Optional key that guarantees this request runs at most once. If the key was already used by your project, the request is refused with `409` instead of being performed a second time. Keys are scoped to your project, expire after 24 hours (configurable when self-hosting), and must be 1-255 printable ASCII characters."
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request body failed schema validation. `error.errors` lists the offending fields.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Request validation failed",
                "statusCode": 422,
                "requestId": "8f14e45f-ceea-467a-9575-1f0f38e0b1c2",
                "errors": [
                  {
                    "field": "to",
                    "message": "Invalid email",
                    "code": "invalid_string"
                  }
                ],
                "suggestion": "Please check the API documentation for the correct request format."
              },
              "timestamp": "2025-01-15T10:30:00.000Z"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INVALID_API_KEY",
                "message": "Invalid secret API key. This endpoint requires a secret key (sk_*), not a public key.",
                "statusCode": 401,
                "requestId": "8f14e45f-ceea-467a-9575-1f0f38e0b1c2"
              },
              "timestamp": "2025-01-15T10:30:00.000Z"
            }
          }
        }
      },
...
      "IdempotencyConflict": {
        "description": "Idempotency-Key already used. The request was refused, not performed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "error": {
                "code": "IDEMPOTENCY_KEY_REUSED",
                "message": "Idempotency-Key \"order-1234-receipt\" has already been used",
                "statusCode": 409,
                "requestId": "8f14e45f-ceea-467a-9575-1f0f38e0b1c2",
                "details": {
                  "key": "order-1234-receipt",
                  "originalRequest": "POST /v1/send",
                  "originalRequestAt": "2025-01-15T10:30:00.000Z",
                  "originalStatusCode": 200
                },
                "suggestion": "This Idempotency-Key was already used, so the request was refused rather than performed twice. Generate a new key for a genuinely new request."
              },
              "timestamp": "2025-01-15T10:31:00.000Z"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/track": {
      "post": {
        "tags": ["Public API"],
        "summary": "Track event",
        "description": "Track an event for a contact. Automatically creates or upserts the contact, then records the event. Tracked events can be used as workflow triggers, segment filters, and audience filters.\n\n**Reserved event names** (rejected with `VALIDATION_ERROR` and code `reserved_event`): anything matching `email.*`, `contact.subscribed`, `contact.unsubscribed`, `segment.<slug>.entry`, `segment.<slug>.exit`. These are emitted by Plunk itself.\n\n**Idempotency**: re-tracking the same event creates a new event record. Send an `Idempotency-Key` header to have a repeated request refused with `409` instead.",
        "operationId": "trackEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "event"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Contact email. The contact is auto-created if it doesn't exist."
                  },
                  "event": {
                    "type": "string",
                    "description": "Event name. Cannot match the reserved patterns above."
                  },
                  "subscribed": {
                    "type": "boolean",
                    "description": "Subscription state to apply to the contact. **New** contacts default to subscribed (`true`). **Existing** contacts keep their current state unless you pass an explicit value here. Pass `false` to track an event without resubscribing an unsubscribed contact."
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Contact data and one-off event variables. Persistent values (primitives, plain objects) are saved on the contact and become available as template variables. Pass `{ value, persistent: false }` for one-shot variables that should not be stored on the contact (e.g. order IDs, transaction details). `null` deletes a field. Empty strings are ignored. Reserved keys are filtered out — see the contacts concept page."
                  }
                }
              },
              "example": {
                "email": "user@example.com",
                "event": "purchase",
                "data": {
                  "product": "Premium Plan",
                  "amount": 99
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event tracked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "contact": {
                          "type": "string",
                          "description": "Contact ID"
                        },
                        "event": {
                          "type": "string",
                          "description": "Event ID"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ]
      }
    },
...
  },
...
}